+++ src/index.js
@@ -31,8 +31,8 @@ const ReduxStoreContext = createContext(warningObject);
 
 // helper hooks
 
-const forcedReducer = state => !state;
-const useForceUpdate = () => useReducer(forcedReducer, false)[1];
+const forcedReducer = state => state + 1;
+const useForceUpdate = () => useReducer(forcedReducer, 0)[1];
 
 const useProxyfied = (state) => {
   // cache