+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Remove experimental useReduxStateMapped
 ### Changed
 - useLayoutEffect and keep latest state after update (see #20)
+- useForceUpdate uses counter instead of boolean (see #20)
 
 ## [1.8.0] - 2019-04-02
 ### Changed