+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 ### Changed
 - useLayoutEffect and keep latest state after update (see #20)
 - useForceUpdate uses counter instead of boolean (see #20)
+- update proxyequal v2.0.5
 
 ## [1.8.0] - 2019-04-02
 ### Changed